Detect whether we are running on RHEL5 or not
Making packages for RHEL5 requires changes to the .spec files
themselves that should only be done if we actually are building
on this platform. This encapsulates the test into a simple-to-use
predicate. (And instead of having the test inside every script,
can be put in front of them in the main script).